feat(room): implement join request management and refactor DM handling
- Added `DmTransactions` to streamline direct message creation and lookup.
- Introduced `JoinRequestService` and related classes for managing join requests.
- Updated `RoomService` and `RoomController` with enhanced member management functionality.
- Enhanced `RoomRepository` and `RoomMemberRepository` to support new query methods for DM and member roles.
- Updated database schema to include constraints for DMs and support join requests.

+++ src/main/java/io/github/nujanzh/yotsubato/dto/room/RoomPreview.java
@@ -4,5 +4,5 @@ import io.github.nujanzh.yotsubato.model.room.RoomType;
 
 import java.util.UUID;
 
-public record RoomPreview(UUID id, String name, RoomType type, String description, int memberCount)
+public record RoomPreview(UUID id, String name, RoomType type, String description, long memberCount)
         implements RoomResponse {}
